Extract LLMExecutionAgent from DocsSummarizer
Made-with: Cursor
