+++ .gitignore
@@ -18,6 +18,7 @@ scratch
 
 # ignore locally significant directories
 tmp
+local_testing/
 data
 vector_db
 vendor
